+++ src/main/java/com/bootcamp/demo/controller/FirebaseController.java
@@ -1,4 +1,4 @@
-package com.bootcamp.demo.Controller;
+package com.bootcamp.demo.controller;
 
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.cloud.firestore.CollectionReference;
