+++ src/main/java/com/bootcamp/demo/exception/ServiceException.java
@@ -1,4 +1,4 @@
-package com.bootcamp.demo.Exceptions;
+package com.bootcamp.demo.exception;
 
 public class ServiceException extends Exception{
     public ServiceException(String message) {
